Update doc, enhance test_report.py and move tempdir fixture to the root
