+++ tests/unit/test_report.py
@@ -1,20 +1,14 @@
 # coding: utf-8
-import tempfile
 from logging import getLogger
-from pathlib import Path
 
 from nxdrive.manager import Manager
-from nxdrive.options import Options
 from nxdrive.report import Report
 
 
-@Options.mock()
-def test_logs():
+def test_logs(tempdir):
     log = getLogger(__name__)
-    folder = Path(tempfile.mkdtemp("-nxdrive-tests"))
-    Options.nxdrive_home = folder
 
-    with Manager() as manager:
+    with Manager(home=tempdir) as manager:
         log.debug("Strange encoding \xe8 \xe9")
 
         # Crafted problematic logRecord
@@ -26,5 +20,15 @@ def test_logs():
             log.exception(str(e))
             log.exception(e)
 
-        report = Report(manager, folder / "report")
+        # Test raw report (calling the Report class manually)
+        report = Report(manager)
         report.generate()
+        path = report.get_path()
+        assert path.is_file()
+        assert path.suffix == ".zip"
+
+        # Test the report managed by the Manager
+        path_managed = manager.generate_report()
+        assert path_managed.is_file()
+        assert path_managed.suffix == ".zip"
+        assert path == path_managed
